chore: separate out firestore repository transaction it tests.
