+++ spring-cloud-gcp-data-firestore/src/test/java/com/google/cloud/spring/data/firestore/it/FirestoreRepositoryIntegrationTests.java
@@ -17,17 +17,12 @@
 package com.google.cloud.spring.data.firestore.it;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.mockito.ArgumentMatchers.any;
-import static org.mockito.Mockito.reset;
-import static org.mockito.Mockito.times;
-import static org.mockito.Mockito.verify;
 
 import com.google.cloud.Timestamp;
 import com.google.cloud.spring.data.firestore.SimpleFirestoreReactiveRepository;
 import com.google.cloud.spring.data.firestore.entities.User;
 import com.google.cloud.spring.data.firestore.entities.User.Address;
 import com.google.cloud.spring.data.firestore.entities.UserRepository;
-import com.google.cloud.spring.data.firestore.transaction.ReactiveFirestoreTransactionManager;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
@@ -44,35 +39,25 @@ import org.springframework.data.domain.Sort.Order;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.util.AopTestUtils;
-import org.springframework.transaction.reactive.TransactionalOperator;
-import org.springframework.transaction.support.DefaultTransactionDefinition;
 import reactor.core.publisher.Flux;
 import reactor.core.publisher.Mono;
 import reactor.test.StepVerifier;
 
 @EnabledIfSystemProperty(named = "it.firestore", matches = "true")
 @ExtendWith(SpringExtension.class)
-@ContextConfiguration(classes = FirestoreTransactionIntegrationTestsConfiguration.class)
-@DisabledInNativeImage
+@ContextConfiguration(classes = FirestoreIntegrationTestsConfiguration.class)
 class FirestoreRepositoryIntegrationTests {
   // tag::autowire[]
   @Autowired UserRepository userRepository;
   // end::autowire[]
 
-  // tag::autowire_tx_manager[]
-  @Autowired ReactiveFirestoreTransactionManager txManager;
-  // end::autowire_tx_manager[]
-
   // tag::autowire_user_service[]
   @Autowired UserService userService;
   // end::autowire_user_service[]
 
-  @Autowired ReactiveFirestoreTransactionManager transactionManager;
-
   @BeforeEach
   void cleanTestEnvironment() {
     this.userRepository.deleteAll().block();
-    reset(this.transactionManager);
   }
 
   @Test
@@ -118,41 +103,6 @@ class FirestoreRepositoryIntegrationTests {
         .verifyComplete();
   }
   // end::repository_built_in[]
-
-  @Test
-  void transactionalOperatorTest() {
-    // tag::repository_transactional_operator[]
-    DefaultTransactionDefinition transactionDefinition = new DefaultTransactionDefinition();
-    transactionDefinition.setReadOnly(false);
-    TransactionalOperator operator =
-        TransactionalOperator.create(this.txManager, transactionDefinition);
-    // end::repository_transactional_operator[]
-
-    // tag::repository_operations_in_a_transaction[]
-    User alice = new User("Alice", 29);
-    User bob = new User("Bob", 60);
-
-    this.userRepository
-        .save(alice)
-        .then(this.userRepository.save(bob))
-        .as(operator::transactional)
-        .block();
-
-    this.userRepository
-        .findAll()
-        .flatMap(
-            a -> {
-              a.setAge(a.getAge() - 1);
-              return this.userRepository.save(a);
-            })
-        .as(operator::transactional)
-        .collectList()
-        .block();
-
-    assertThat(this.userRepository.findAll().map(User::getAge).collectList().block())
-        .containsExactlyInAnyOrder(28, 59);
-    // end::repository_operations_in_a_transaction[]
-  }
 
   @Test
   // tag::repository_part_tree[]
@@ -359,49 +309,6 @@ class FirestoreRepositoryIntegrationTests {
             .block();
 
     assertThat(pagedUsers).containsExactlyInAnyOrder("Cloud");
-  }
-
-  @Test
-  void declarativeTransactionRollbackTest() {
-    this.userService.deleteUsers().onErrorResume(throwable -> Mono.empty()).block();
-
-    verify(this.transactionManager, times(0)).commit(any());
-    verify(this.transactionManager, times(1)).rollback(any());
-    verify(this.transactionManager, times(1)).getReactiveTransaction(any());
-  }
-
-  @Test
-  void declarativeTransactionCommitTest() {
-    User alice = new User("Alice", 29);
-    User bob = new User("Bob", 60);
-
-    this.userRepository.save(alice).then(this.userRepository.save(bob)).block();
-
-    this.userService.updateUsers().block();
-
-    verify(this.transactionManager, times(1)).commit(any());
-    verify(this.transactionManager, times(0)).rollback(any());
-    verify(this.transactionManager, times(1)).getReactiveTransaction(any());
-
-    assertThat(this.userRepository.findAll().map(User::getAge).collectList().block())
-        .containsExactlyInAnyOrder(28, 59);
-  }
-
-  @Test
-  void transactionPropagationTest() {
-    User alice = new User("Alice", 29);
-    User bob = new User("Bob", 60);
-
-    this.userRepository.save(alice).then(this.userRepository.save(bob)).block();
-
-    this.userService.updateUsersTransactionPropagation().block();
-
-    verify(this.transactionManager, times(1)).commit(any());
-    verify(this.transactionManager, times(0)).rollback(any());
-    verify(this.transactionManager, times(1)).getReactiveTransaction(any());
-
-    assertThat(this.userRepository.findAll().map(User::getAge).collectList().block())
-        .containsExactlyInAnyOrder(28, 59);
   }
 
   @Test
